+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- XML parser now preserves comments and processing instructions during round-trip (#175).
+
 ### Changed
 - Swapped to use goccy/go-json for improved performance. Thanks @imix
 - Updated model `IsScalar` internals to improve efficiency. Thanks @imix